+++ CHANGELOG.md
@@ -1,4 +1,13 @@
+## 0.2.0
+
+- **New**
+  - `JoJo's Menacing: Screen Size Animation`
+    - Added a new animation based on JoJo's Menacing Animation
+  - `End Credits`
+    - Added a new end credits animation that can be passed an array of lines or a very lon string with new lines in it to display a movie end credits animation
+
 ## 0.1.0
+
 - **New**
   - `Demon Slayer Transition`
     - Added a new animation based on Demon Slayer's Transition screens they do